+++ src/main/java/com/kodilla/ecommercee/cart/controller/CartController.java
@@ -1,7 +1,7 @@
 package com.kodilla.ecommercee.cart.controller;
 
-import com.kodilla.ecommercee.cart.domain.CartDto;
-import com.kodilla.ecommercee.cart.domain.UserDto;
+import com.kodilla.ecommercee.cart.dto.CartDto;
+import com.kodilla.ecommercee.cart.dto.UserDto;
 import com.kodilla.ecommercee.product.dto.ProductDto;
 import org.springframework.web.bind.annotation.*;
