+++ src/main/java/com/kodilla/ecommercee/cart/domain/Cart.java
@@ -1,10 +1,10 @@
 package com.kodilla.ecommercee.cart.domain;
 
+import com.kodilla.ecommercee.cart.dto.UserDto;
 import com.kodilla.ecommercee.product.dto.ProductDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
