+++ src/main/java/com/kodilla/ecommercee/generic/GenericEntity.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.generic;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
