+++ src/main/java/com/kodilla/ecommercee/group/controller/GroupController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.group.controller;
+
+public class GroupController {
+}
