+++ src/main/java/com/kodilla/ecommercee/product/dto/ProductDto.java
@@ -3,7 +3,6 @@ package com.kodilla.ecommercee.product.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @NoArgsConstructor
 @AllArgsConstructor
