+++ src/main/java/com/kodilla/ecommercee/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.user.controller;
+
+public class UserController {
+}
