+++ src/test/java/com/kodilla/ecommercee/cart/controller/CartControllerTest.java
@@ -1,8 +1,7 @@
 package com.kodilla.ecommercee.cart.controller;
 
-import com.kodilla.ecommercee.cart.domain.CartDto;
-import com.kodilla.ecommercee.cart.domain.UserDto;
-import com.kodilla.ecommercee.product.controller.ProductController;
+import com.kodilla.ecommercee.cart.dto.CartDto;
+import com.kodilla.ecommercee.cart.dto.UserDto;
 import com.kodilla.ecommercee.product.dto.ProductDto;
 import org.junit.Test;
