+++ src/test/java/com/kodilla/ecommercee/cart/domain/CartDtoTest.java
@@ -1,5 +1,6 @@
 package com.kodilla.ecommercee.cart.domain;
 
+import com.kodilla.ecommercee.cart.dto.CartDto;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
